Changed Textures
Since the stock Doom 2 Brown textures were removed, it made a lot of maps see those as unknown textures.
For the forts in UT'99 part, Brownhug from the default Doom 2 Textures was used. Looks a bit scuffed, but that's the drawback of Doom/Hexen Format maps.
Technoir needs a lot more, but it must be done in a way so it doesn't look ugly, since the brown textures were also used on the ceilings.