Small change to the first three XMAS maps
In lieu of Burrowsville including almost every power-up box except the first three shots, I changed it so they show up in the first three maps (Slicer is moved into the GBA secret, Seeker is replaced with Spread Gun and the Slicer in the first map is replaced with the upgraded LFG). I didn't touch the ones in Icebox, since it's a secret level and it's possible some might miss them and won't have them until Burrowsville.